+++ src/matchbox/server/postgresql/utils/query.py
@@ -54,7 +54,7 @@ def _get_dataset_resolution(
 def _resolve_thresholds(
     lineage_truths: dict[str, float],
     resolution: Resolutions,
-    threshold: float | None,
+    threshold: int | None,
 ) -> dict[int, float]:
     """Resolves final thresholds for each resolution in the lineage based on user input.
 
@@ -90,7 +90,7 @@ def _resolve_thresholds(
     return resolved_thresholds
 
 
-def _get_valid_clusters_for_resolution(resolution_id: int, threshold: float) -> Select:
+def _get_valid_clusters_for_resolution(resolution_id: int, threshold: int) -> Select:
     """Get clusters that meet the threshold for a specific resolution."""
     return select(Probabilities.cluster.label("cluster")).where(
         and_(
@@ -135,7 +135,7 @@ def _resolve_cluster_hierarchy(
     dataset_id: int,
     resolution: Resolutions,
     engine: Engine,
-    threshold: float | None = None,
+    threshold: int | None = None,
 ) -> Select:
     """Resolves the final cluster assignments for all records in a dataset.
 
@@ -257,7 +257,7 @@ def query(
     engine: Engine,
     source_address: SourceAddress,
     resolution_id: int | None = None,
-    threshold: float | None = None,
+    threshold: int | None = None,
     limit: int = None,
 ) -> pa.Table:
     """Queries Matchbox and the Source warehouse to retrieve linked data.
@@ -485,7 +485,7 @@ def _build_match_query(
     source_resolution_id: int,
     resolution: str,
     session: Session,
-    threshold: float | None = None,
+    threshold: int | None = None,
 ) -> Select:
     """Builds the SQL query that powers the match function."""
     # Get truth resolution
@@ -533,7 +533,7 @@ def match(
     target: SourceAddress | list[SourceAddress],
     resolution: str,
     engine: Engine,
-    threshold: float | None = None,
+    threshold: int | None = None,
 ) -> Match | list[Match]:
     """Matches an ID in a source dataset and returns the keys in the targets.
